Add files via upload
The document explores the role of human cybersecurity in the socio-technical system, the role of human in cybersecrity and risks posed by the gap in human in adressing  cybersecurity  defense system. 
Additionally, the document explores the models and frameworks designed to better understand the huan cybersecurity and concludes by examining the adoption strategies to better enhance humans in cybersecurity.