Add files via upload
This report presents a comprehensive analysis of an unidentified malware in the network infrastructure of Cleeve Corporation. Employing a combination of static and dynamic analysis, this paper aims to identify and analyse the behaviour of the malware, including the malware’s method of infection, propagation and Convert operation, also discussing the potential damage that could be caused by the presence of the malware in the system and countermeasures. This paper further analyses emerging technologies that enhance malware detection and develop security resilience against evolving threats.